ss, rss, drss return LinearIOSystem
